PMI-753: Move `get_pygsheets_client()` to a new module `gsheets` and make it throw exception if a required environment variable doesn't exist.
