+++ performance/config.py
@@ -22,6 +22,8 @@ class Config:
     PIWIK_LIMIT = '-1'
     PIWIK_BASE_URL = 'https://analytics-hub-prod-a-dmz.ida.digital.cabinet-office.gov.uk/index.php'
     DEFAULT_OUTPUT_PATH = os.path.join(BASE_DIR, 'output')
+    # This is only used if Google auth credentials aren't already present in environment variables See
+    # `performance.gsheets.get_pygsheets_client` for implementation details.
     GSHEETS_CREDENTIALS_FILE = os.path.join(
         VERIFY_DATA_PIPELINE_CONFIG_PATH, 'credentials', 'google_sheets_credentials.json')
